feat: move OrderDetail from views to components
